arm: topology: Define TC2 energy and provide it to the scheduler
!!! This patch is only here to be able to test provisioning of energy
related data from an arch topology shim layer to the scheduler.
Since there is no code today which deals with extracting energy related
data from the dtb or acpi, and process it in the topology shim
layer, the content of the sched_group_energy structures as well as the
idle_state and capacity_state array are hard-coded here !!!

This patch defines the sched_group_energy structure as well as the
idle_state and capacity_state array for the cluster (relates to sched
groups (sg's) in DIE sched domain (sd) level) and for the core (relates
to sg's in MC sd level) for a Cortex A7 as well as for a Cortex A15.
It further provides related implementations of the sched_domain_energy_f
functions (cpu_cluster_energy() and cpu_core_energy()).

To be able to propagate this information from the topology shim layer to
the scheduler, the elements of the arm_topology[] table have been
provisioned with the appropriate sched_domain_energy_f functions.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>