sched: Introduce 'int cpu' parameter for sched_domain_energy_f
The folding of two adjacent sd levels via sd degenerate doesn't work for
all sd levels.  E.g. it is not possible to use this feature to provide
per-cpu sd energy in sd level DIE (former CPU) on TC2.  This patch adds
an 'int cpu' parameter to the sched_domain_energy_f function.

It was discussed on LKML that the folding of sd levels approach is
preferable over the cpu parameter approach, simply because the user (the
arch specifying the sd topology table) can introduce less errors.  But
since it is not working the int cpu parameter is the only way out.  It's
possible to use the folding of sd levels approach for
sched_domain_flags_f and the cpu parameter approach for the
sched_domain_energy_f though.

This patch also provides the check function check_sd_energy_data() which
throws a BUG_ON() if there are inconsistencies in the provisioning of sd
energy data inside a sched group.  This check function becomes necessary
now that sched_domain_energy_f has a cpu parameter.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>

Conflicts:

	kernel/sched/core.c

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>