sched: energy_diff functions
Introduces energy_diff_load() which finds the energy impacts of
adding or removing load from a specific cpu. The calculation is
based on the energy information provided by the platform through
sched_energy data in the sched_domain hierarchy.

Task utilization is currently based on load_avg_contrib which is
actually load, nor utilization. We don't have a solution for
utilization yet. The same goes for the cpu utilization, where
weighted_cpuload() is used.

Energy impact is based on utilization for now. Wake-ups will be
added later.