sched: Take task wakeups into account in energy estimates
The energy cost of waking a cpu and sending it back to sleep can be
quite significant for short running frequently waking tasks if placed
on an idle cpu in a deep sleep state. By factoring task wakeups in
such tasks can be placed on cpus where the wakeup energy cost is
lower. For example, partly utilized cpus in a shallower idle state,
or cpus in a cluster/die that is already awake.

Current cpu utilization of the target cpu is factored in guess how many
task wakeups that translate into cpu wakeups (idle exits). It is a very
naive approach, but it is virtually impossible to get an accurate estimate.

wake_energy(task) = unused_util(cpu) * wakeups(task) * wakeup_energy(cpu)

There is no per cpu wakeup tracking, so we can't estimate the energy
savings when removing tasks from a cpu. It is also nearly impossible to
figure out which task is the cause of cpu wakeups if multiple tasks are
scheduled on the same cpu.

Support for multiple idle-states per sched_group (e.g. WFI and core
shutdown on ARM) is not implemented yet. wakeup_energy in stuct
sched_energy needs to be a table instead and cpuidle needs to tells
what the most likely state is.