sched: Use energy to guide wakeup task placement
Attempt to pick most energy efficient wakeup in find_idlest_{group, cpu}().
Finding the optimum target requires an exhaustive search through all
cpus in the groups. Instead, the target group is determined based on load
and probing the energy cost on a single cpu in each group. The target cpu
is the cpu with the lowest energy cost.

Clean-up is needed. Split find_target_group() into idlest and lowenergy
to avoid horrible ifdefs?