sched: Disable wake_affine to broaden the scope of wakeup target cpus
SD_WAKE_AFFINE is currently set by default on all levels which means
that wakeups are always handled inside the lowest level sched_domain.
That means, a tiny periodic task is very likely to stay on the cpu it
was forked on forever. To save energy, we need to revisit the task
placement decision every now and again to ensure that we don't keep
waking the same cpu if there are cheaper alternatives.

One way is to simply disable wake_affine and rely on the fork/exec
balancing mechanism (find_idlest_{group, cpu}). This is what this
patch does.

An alternative is to let the platform remove the SD_WAKE_AFFINE flag
from lower levels to increase the search space for select_idle_sibling().