sched, energy: Add idle power contribution to energy model
Idle power is modelled as fraction time that the power domain is
estimated to be unutilized. The actual idle time reported by cpuidle
might differ.

The idle power estimate currently only supports a single idle state
per power (sub-)domain. Extending the support to multiple states
requires a way of predicting which states is going to be the most
likely. This prediction could be provided by cpuidle.