thermal: scpi: Add support for SCPI based thermal sensor
SCPI provides an interface to -
* query the sensors in the system
* query the value of these sensors

Extend the SCPI protocol implementation to support these sensors.

Building on these SCPI commands, integrate the thermal sensor
"SENSOR_TEMP_SOC" with the linux thermal framework and create a SoC
thermal zone using the of-thermal APIs.

Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>

This is the changes to the scpi protocol split out from Punit's
original commit.

Signed-off-by: Jon Medhurst <tixy@linaro.org>