sched: Use energy model in load balance path
Attempt to pick the source cpu which potentially gives the maximal energy
savings in case the min(amount of utilization which the destination cpu is
additionally able to handle, the current utilization on the source cpu) is
taken away from it.  Finding the optimum source requires an exhaustive
search through all cpus in the groups.  Instead, the source group is
determined based on utilization and probing the energy cost on a single
cpu in each group.

This implementation is not providing an actual energy aware load
balancing right now. It is only trying to showcase the way to find the
most suitable source queue (cpu) based on the energy aware data. The
actual load balance is still done based on the calculated load based
imbalance.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>