sched: Likely idle state statistics placeholder
The scheduler is currently completely unaware of idle-states cpus. To
make informed decisions using the sched_group_energy idle_states list it
is necesarry to know which idle-state a cpu (or group of cpus) is most
likely to be in when it is idle.

For example when migrating a task that wakes up periodically, the wakeup
energy expense depends on the idle-state the destination cpu is most
likely to be in when idle.

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>