cpuidle: Add a simple irq governor
This simple governor takes into account the predictable events: the timer sleep
duration and the next expected irq sleep duration. By mixing both it deduced
what idle state fits better.

The main purpose of this governor is to handle the guessed next events in a
categorized way:
	1. deterministic events : timers
	2. guessed events : IOs
	3. predictable events : keystroke, incoming network packet, ...

This governor is aimed to be moved later near the scheduler, so this one
can inspect/inject more informations and act proactively rather than
reactively.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>

Conflicts:
	drivers/cpuidle/Kconfig