arm64: Juno: Comment out the secondary HDLCD controller
The HDLCD controllers on Juno share the same PLL for the pixel
clock output. If the secondary output is not connected to a
monitor with the same resolution as the primary output, or
to a monitor with VGA resolution, it will cause the pixel
clock to be set to the wrong value for the primary output.
Comment out the secondary HDLCD controller for now.

Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
[Modified for newer device-tree]
Signed-off-by: Jon Medhurst <tixy@linaro.org>