sched: Architecture function for current and average compute capacity
The scheduler and particularly the load tracking statistics need to
track changes in the cpu compute capacities (due to DVFS) to compensate
for any changes and provide scale invariant load tracking.

Two funtions with architecture specific implementations are needed for
this purpose: arch_scale_curr_capacity(cpu) that returns the current
compute capacity (or best guestimate that the architecture can provide),
and arch_scale_avg_capacity(task) that returns the average compute
capacity of the cpu on which the task was running since last call to
this function. The latter is needed for scale-invariant task
load-tracking. It is assumed that arch_scale_avg_capacity() is always
called before task migration so the function implementation doesn't have
to deal with cross-cpu capacity averages.

Each architecture is free to use any special architecture capabilities
(such as performance counters) for their implementations. Average
capacity will most likely require some state (counter readings for
example) to be save for each task. This could be stored in the
architecture specific thread_struct which can be reached from the
task_struct pointer provided in the function call.