arm: TC2 specific implemetations of compute capacity scaling functions
This is a simple first stab at implementing the architecture specific
functions needed for the scale-invariant load-tracking in the scheduler.
Infrastructure for capacity/frequency states information is missing, so
everything is hard-coded in the architecture specific code for now. That
includes duplicating information that is already in cpufreq. The aim for
now is just a proof of concept that avoids architecture specific hacks
in generic code in the scheduler and cpufreq.