wip/sched: Remove if condition in energy_diff_util
Remove the if condition 'if (curr_cap_idx < 0 || !(sd->flags &
SD_SHARE_CAP_STATES))' inside for_each_domain(cpu, sd). This if condition
is always true on TC2, curr_cap_idx is -1 for MC level and
SD_SHARE_CAP_STATES is not set for DIE (or SYS) level.

If accepted this patch has to be squashed into:

  f15fc6d610de - sched: Energy model functions

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>