wip/sched: Rework init_sched_energy
We have the problem that we can't express system wide energy by attaching
sched_energy onto sched_group which itself is attached to a sched domain.
The issue is that the scheduler does not allow sched domains with only one
group since from the scheduling perspective this would be rather useless.

So far we had a special system sched energy structure (sse) which is
initialized when the arch specific topology table provides an energy
function cpu_sys_energy in its last entry, the one which already has its
mask function set to NULL. The problem with this is that we have to treat
the SYS sched domain level differently then all the others since we can't
iterate over its sched domain pointer simply because it doesn't exist.

This patch creates not only a system sched energy structure but the whole
set consisting of a sched domain (ssd), a sched group and sched energy.
It still uses the approach to get its energy data via cpu_sys_energy.

The cpumask of the system sched group is taken from the sched domain mask
of the highest (non-phantom) sched domain automatically. The name of the
system sched domain is set to SYS in the arch specific topology table for
debugging purposes.

The ssd is still not a per-cpu variable exactly like sse was. I can't see
any issues with that since we only read the system ssd->sg->sge data.

This patch also provides way to iterate over all sched domains inclusive
ssd in case it exists. This should be used by all sched energy related
code such as energy_diff_util.

If accepted this patch has to be squashed into:

  8d674715d475 - sched: Introduce system-wide sched_energy
  f15fc6d610de - sched: Energy model functions

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>