wip/sched: Refactor energy_diff_util
In it's current form, energy_diff_util is very hard to understand :-)
Try to separate the calculation of the affected util pair and the spare
util pair into the new function calc_diff_util. This one deals with all
the differences of the sd's (i.e. !sd->child and sd->flags &
SD_SHARE_CAP_STATES).
Re-use the idea of struct lb_env to be able to pass a lot of arguments
and to retrieve multiple return values from this function.
The function find_max_util becomes redundant then.

If accepted this patch has to be squashed into:

  f15fc6d610de - sched: Energy model functions

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>