HMP: Make idle pull optional
Not all platforms can benefit from idle pull replacing
forced migration. Specifically, if the exit latency of
the big CPUs is less than a few milliseconds there is
no benefit to pulling compared to forced migration.

This patch allows the idle pull migration method to be
switched off at runtime so that platforms where this
is not useful can disable it when setting other power
management options.