sched: Introduce scale-invariant load tracking
The per-entity load-tracking is currently unaware of any variations in
cpu compute capacity due to frequency scaling (cpufreq) or task
migrations between cpus with different micro-architectures (e.g. ARM
big.LITTLE). Comparing tracked loads between different cpus might
therefore be quite misleading.

This patch introduces a scale-invariance scaling factor to the
load-tracking computation that can be used to compensate for compute
capacity variations. The scaling factor is to be provided by the
architecture through an arch specific function. It may be as simple as:

	current_freq(cpu) * 1024 / max_freq(cpu)

If the architecture has more sophisticated ways of tracking compute
capacity, they can do so in the their implementation. By default, no
scaling is applied.

The patch is loosely based on a patch by Chris Redpath
<Chris.Redpath@arm.com>.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>