FROMLIST: sched: Track blocked utilization contributions
Introduces the blocked utilization, the utilization counter-part to
cfs_rq->blocked_load_avg. It is the sum of sched_entity utilization
contributions of entities that were recently on the cfs_rq and are
currently blocked. Combined with the sum of utilization of entities
currently on the cfs_rq or currently running
(cfs_rq->utilization_load_avg) this provides a more stable average
view of the cpu usage.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

(am from https://patchwork.kernel.org/patch/6391041)
Signed-off-by: Juri Lelli <juri.lelli@arm.com>

BUG=chromium:48849
TEST=build and boot tested

Change-Id: I1526a1acd2ad7b35e01fd4630da4a763decbd377