cpuidle: Factor out the selection loop of the idle state
The selection loop follows the logic "choose an idle state fulfilling the sleep
duration and the exit latency constraints". It can be encapsulated into a
function and reused in different places. It gives the API to choose an idle
state regarding the timing information passed as parameter.

The 'cpuidle_find_deepest_state' function does the similar selection. The new
function can be used instead with infinite time constraints as we are going to
suspend.

That is one more step to the cpuidle/scheduler integration.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>